pyproject.toml
[tool.poetry]
name = "william-travel-backend"
version = "0.1.0"
description = "Backend API for William Travel application"
authors = ["William Nguyen"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
sqlalchemy = "^2.0.21"
asyncpg = "^0.28.0"
psycopg2-binary = "^2.9.9"
alembic = "^1.12.0"
python-dotenv = "^1.0.0"
geoalchemy2 = "^0.14.0"
gunicorn = "^21.2.0"
shapely = "^2.0.1"
email-validator = "^2.0.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-asyncio = "^0.21.1"
httpx = "^0.25.0"
black = "^23.9.1"
isort = "^5.12.0"
mypy = "^1.5.1"
ruff = "^0.0.292"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
